refactor: extract cloud session fetch/import logic into cloud-sessions.ts
